Fixed project structure.
