--- Katas.BinaryGap/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="NUnit" version="3.4.1" targetFramework="net452" />
-</packages>
