+++ test/nunit/BinaryGap/TestsBinaryGap.cs
