+++ test/nunit/Equi/Equi.cs
