+++ test/nunit/Equi/TestEqui.cs
